refactor: move types around to be more localised, no need for common types
