+++ src/client/functions.ts
@@ -1,11 +1,10 @@
-import { NuiCallbackFunc } from "../common/types";
-
-type CallbackFn<T> = (data: T, cb: NuiCallbackFunc) => void;
+import { CallbackFn } from "./types";
 
 /**
  * A wrapper for handling NUI Callbacks
- *  @param event - The event name to listen for
- *  @param callback - The callback function
+ * @typeParam T The data type of the data sent back to the client
+ * @param event - The event name to listen for
+ * @param callback - The callback function
  */
 export const RegisterNuiCB = <T = any>(event: string, callback: CallbackFn<T>): void => {
   RegisterNuiCallbackType(event);
