+++ src/client/index.ts
@@ -1,3 +1,3 @@
 export * from "./functions";
 export * from "./cl_utils";
-export * from "./cl_logger";
+export * from "./types";
