+++ src/index.ts
@@ -1,4 +1,3 @@
 export * from "./client";
 export * from "./server";
 export * from "./common/helpers";
-export * from "./common/types";
